requirements.txt
alembic==0.9.9
aniso8601==3.0.0
astroid==1.6.3
attrs==17.4.0
click==6.7
colorama==0.3.9
coverage==4.5.1
flasgger==0.8.3
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
limits==1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.1.0
nose==1.3.7
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
PyJWT==1.6.1
pylint==1.8.4
pytest==3.5.1
pytest-cov==2.5.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
wrapt==1.10.11
